requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
plotly==5.24.1
jinja2==3.1.4
faiss-cpu==1.9.0
google-generativeai==0.8.3
python-dotenv==1.0.1
camelot-py==0.11.0
pdfplumber==0.11.4
rapidfuzz==3.10.1
PyMuPDF==1.24.9
pdf2image==1.17.0
pytesseract==0.3.13
tabula-py==2.9.0
umap-learn==0.5.6
